+++ playground/backend/app/controllers/health_controller.ts
@@ -1,0 +1,10 @@
+import { healthChecks } from '#start/health'
+import { handler } from '@folie/castle/helpers'
+
+export default class Controller {
+  handle = handler(async () => {
+    const report = await healthChecks.run()
+
+    return { report }
+  })
+}