+++ playground/backend/start/routes.ts
@@ -56,38 +56,42 @@ router
 
         router
           .group(() => {
-            router.get('', [() => import('#controllers/note/list_controller')])
+            router
+              .group(() => {
+                router.get('', [() => import('#controllers/note/list_controller')])
 
-            router.get(':noteId', [() => import('#controllers/note/show_controller')])
+                router.get(':noteId', [() => import('#controllers/note/show_controller')])
 
-            router.post('', [() => import('#controllers/note/create_controller')])
+                router.post('', [() => import('#controllers/note/create_controller')])
 
-            router.put(':noteId', [() => import('#controllers/note/update_controller')])
+                router.put(':noteId', [() => import('#controllers/note/update_controller')])
 
-            router.delete(':noteId', [() => import('#controllers/note/delete_controller')])
+                router.delete(':noteId', [() => import('#controllers/note/delete_controller')])
+
+                router
+                  .group(() => {
+                    router.put(':noteId', [() => import('#controllers/note/tag/update_controller')])
+                  })
+                  .prefix('tag')
+              })
+              .prefix('note')
 
             router
               .group(() => {
-                router.put(':noteId', [() => import('#controllers/note/tag/update_controller')])
+                router.get('', [() => import('#controllers/tag/list_controller')])
+
+                router.get(':tagId', [() => import('#controllers/tag/show_controller')])
+
+                router.post('', [() => import('#controllers/tag/create_controller')])
+
+                router.put(':tagId', [() => import('#controllers/tag/update_controller')])
+
+                router.delete(':tagId', [() => import('#controllers/tag/delete_controller')])
               })
               .prefix('tag')
+
+            router.get('health', [() => import('#controllers/health_controller')])
           })
-          .prefix('note')
-          .use(middleware.auth())
-
-        router
-          .group(() => {
-            router.get('', [() => import('#controllers/tag/list_controller')])
-
-            router.get(':tagId', [() => import('#controllers/tag/show_controller')])
-
-            router.post('', [() => import('#controllers/tag/create_controller')])
-
-            router.put(':tagId', [() => import('#controllers/tag/update_controller')])
-
-            router.delete(':tagId', [() => import('#controllers/tag/delete_controller')])
-          })
-          .prefix('tag')
           .use(middleware.auth())
 
         router.get('ping', [() => import('#controllers/ping_controller')])