+++ packages/blueprint_lib/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @folie/blueprint-lib
 
+## 0.0.11
+
+### Patch Changes
+
+- Refactor types to use type-fest utilities
+
 ## 0.0.10
 
 ### Patch Changes